Lots of small tweaks, a little code cleanup, MQTT improvements, threshold warnings and more.
Generally cleaned up the code and log messages.
Redid some of the MQTT setup code and improved it a little. Errors regarding the constant reconnect to MQTT seems to have been fixed (at least on my end).
Added a threshold warning in the navbar that triggers when the chosen data column enters the threshold area. Note that it only triggers on new MQTT messages and it will always display the newest occurance of the threshold warning.
Added a new setting to set the threshold target data column.
Added MQTT broker settings to the "config.json" file for quickly changing MQTT broker.
Mostly fixed graph zoom issues by adding a new way of zooming. There is now a zoom bar at the bottom of each graph that hopefully should make it easier to zoom. Zooming now also doesn't filter away values outside of the graph as before.
Also added a way to save the current graph as a png image.